refactor: Controller 클래스 인스턴스 생성 책임 분리
